remove pause icon from navbar, move landing page component to component folder and update imports
