+++ src/App.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import LandingPage from './landingPage';
+import LandingPage from './components/landingPage';
 import Navbar from './components/navbar'
 
 class App extends Component {
